Fully implement and test the role APIs

+++ odp/api2/routers/role.py
@@ -2,12 +2,13 @@ from typing import List
 
 from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy import select
-from starlette.status import HTTP_404_NOT_FOUND, HTTP_409_CONFLICT
+from starlette.status import HTTP_404_NOT_FOUND, HTTP_409_CONFLICT, HTTP_403_FORBIDDEN
 
+from odp import ODPScope
 from odp.api2.models import RoleModel, RoleSort
-from odp.api2.routers import Pager, Paging
+from odp.api2.routers import Pager, Paging, Authorize, Authorized
 from odp.db import Session
-from odp.db.models import Role
+from odp.db.models import Role, Scope
 
 router = APIRouter()
 
@@ -18,6 +19,7 @@ router = APIRouter()
 )
 async def list_roles(
         pager: Pager = Depends(Paging(RoleSort)),
+        auth: Authorized = Depends(Authorize(ODPScope.ROLE_READ)),
 ):
     stmt = (
         select(Role).
@@ -25,11 +27,14 @@ async def list_roles(
         offset(pager.skip).
         limit(pager.limit)
     )
+    if auth.provider_ids != '*':
+        stmt = stmt.where(Role.provider_id.in_(auth.provider_ids))
 
     roles = [
         RoleModel(
             id=row.Role.id,
-            name=row.Role.name,
+            scope_ids=[scope.id for scope in row.Role.scopes],
+            provider_id=row.Role.provider_id,
         )
         for row in Session.execute(stmt)
     ]
@@ -43,13 +48,18 @@ async def list_roles(
 )
 async def get_role(
         role_id: str,
+        auth: Authorized = Depends(Authorize(ODPScope.ROLE_READ)),
 ):
     if not (role := Session.get(Role, role_id)):
         raise HTTPException(HTTP_404_NOT_FOUND)
 
+    if auth.provider_ids != '*' and role.provider_id not in auth.provider_ids:
+        raise HTTPException(HTTP_403_FORBIDDEN)
+
     return RoleModel(
         id=role.id,
-        name=role.name,
+        scope_ids=[scope.id for scope in role.scopes],
+        provider_id=role.provider_id,
     )
 
 
@@ -58,13 +68,21 @@ async def get_role(
 )
 async def create_role(
         role_in: RoleModel,
+        auth: Authorized = Depends(Authorize(ODPScope.ROLE_ADMIN)),
 ):
+    if auth.provider_ids != '*' and role_in.provider_id not in auth.provider_ids:
+        raise HTTPException(HTTP_403_FORBIDDEN)
+
     if Session.get(Role, role_in.id):
         raise HTTPException(HTTP_409_CONFLICT)
 
     role = Role(
         id=role_in.id,
-        name=role_in.name,
+        scopes=[
+            Session.get(Scope, scope_id)
+            for scope_id in role_in.scope_ids
+        ],
+        provider_id=role_in.provider_id,
     )
     role.save()
 
@@ -74,11 +92,19 @@ async def create_role(
 )
 async def update_role(
         role_in: RoleModel,
+        auth: Authorized = Depends(Authorize(ODPScope.ROLE_ADMIN)),
 ):
+    if auth.provider_ids != '*' and role_in.provider_id not in auth.provider_ids:
+        raise HTTPException(HTTP_403_FORBIDDEN)
+
     if not (role := Session.get(Role, role_in.id)):
         raise HTTPException(HTTP_404_NOT_FOUND)
 
-    role.name = role_in.name
+    role.scopes = [
+        Session.get(Scope, scope_id)
+        for scope_id in role_in.scope_ids
+    ]
+    role.provider_id = role_in.provider_id,
     role.save()
 
 
@@ -87,8 +113,12 @@ async def update_role(
 )
 async def delete_role(
         role_id: str,
+        auth: Authorized = Depends(Authorize(ODPScope.ROLE_ADMIN)),
 ):
     if not (role := Session.get(Role, role_id)):
         raise HTTPException(HTTP_404_NOT_FOUND)
+
+    if auth.provider_ids != '*' and role.provider_id not in auth.provider_ids:
+        raise HTTPException(HTTP_403_FORBIDDEN)
 
     role.delete()
